refactor(self_metrics): promote to directory, split tests (#1266)
self_metrics.rs (429 LOC) is promoted to a directory module:
- src/self_metrics/mod.rs: 268 LOC (down from 429)
- src/self_metrics/tests.rs: new (~158 LOC)

Pure refactor: zero logic changes, all 9 tests pass.

Closes 4/70 of #1266 (this loop iteration), 12 PRs cumulative since #1285.
